Temporarily disable capa name masking
There is a bug I need to fix when a student re-visits a problem that they
had solved previously without masking enabled, their old state does not
load correclty.

It's simplest to turn off masking now, and re-enable it with the whole fix.

Masking only applies when problems use the new shuffle="true" or answer-pool features,
probably not used outside of Stanford yet. Those problems still work as documented with
this change, they will just not use masking internally.

The change here is a small deleting to remove the masking=True logic, and more
lengthy changes to remove the tests of masking.